+++ examples/go/fill/02_fill.go
@@ -9,21 +9,21 @@ import (
 )
 
 //line "go/fill/02_fill.go":12
-var YYMAXFILL int = 1
+var YYMAXFILL uint = 1
 //line "go/fill/02_fill.re":9
 
-const BUFSIZE int = 4096
+const BUFSIZE uint = 4096
 
 type Input struct {
 	file *os.File
 	buf  []byte
-	cur  int
-	tok  int
-	lim  int
+	cur  uint
+	tok  uint
+	lim  uint
 	eof  bool
 }
 
-func fill(in *Input, need int) int {
+func fill(in *Input, need uint) int {
 	if in.eof { return -1 } // unexpected EOF
 
 	// Error: lexeme too long. In real life can reallocate a larger buffer.
@@ -37,13 +37,13 @@ func fill(in *Input, need int) int {
 
 	// Fill free space at the end of buffer with new data from file.
 	n, _ := in.file.Read(in.buf[in.lim:BUFSIZE])
-	in.lim += n
+	in.lim += uint(n)
 
 	// If read less than expected, this is end of input => add zero padding
 	// so that the lexer can access characters at the end of buffer.
 	if in.lim < BUFSIZE {
 		in.eof = true
-		for i := 0; i < YYMAXFILL; i += 1 { in.buf[in.lim+i] = 0 }
+		for i := uint(0); i < YYMAXFILL; i += 1 { in.buf[in.lim+i] = 0 }
 		in.lim += YYMAXFILL
 	}
 
@@ -139,9 +139,9 @@ func main() () {
 	// Prepare input file: a few times the size of the buffer, containing
 	// strings with zeroes and escaped quotes.
 	f, _ := os.Create(fname)
-	f.WriteString(strings.Repeat(content, BUFSIZE))
+	f.WriteString(strings.Repeat(content, int(BUFSIZE)))
 	f.Seek(0, 0)
-	count := 3 * BUFSIZE // number of quoted strings written to file
+	count := 3 * int(BUFSIZE) // number of quoted strings written to file
 
 	// Prepare lexer state: all offsets are at the end of buffer.
 	// This immediately triggers YYFILL, as the YYLESSTHAN condition is true.
