+++ test/codegen/loop_switch/go/bounds_checking.go
@@ -7,13 +7,13 @@ import (
 	"testing"
 )
 
-var YYMAXFILL int = 1
+var YYMAXFILL uint = 1
 
 
 // Expects YYMAXFILL-padded string.
 func lex(str string) int {
-	var cursor int
-	limit := len(str)
+	var cursor uint
+	limit := uint(len(str))
 	count := 0
 loop:
 	
@@ -112,7 +112,7 @@ yyl:
 
 // Pad string with YYMAXFILL zeroes at the end.
 func pad(str string) string {
-	return str + strings.Repeat("\000", YYMAXFILL)
+	return str + strings.Repeat("\000", int(YYMAXFILL))
 }
 
 func TestLex(t *testing.T) {
