+++ test/golang/003_fill_f.go
@@ -7,19 +7,19 @@ import "fmt"
 import "os"
 
 //line "golang/003_fill_f.go":10
-var YYMAXFILL int = 2
+var YYMAXFILL uint = 2
 //line "golang/003_fill_f.re":7
 
-var SIZE int = 11
+var SIZE uint = 11
 
 type YYCTYPE byte
 type Input struct {
 	file   *os.File
 	data   []byte
-	cursor int
-	marker int
-	token  int
-	limit  int
+	cursor uint
+	marker uint
+	token  uint
+	limit  uint
 	eof    bool
 	state  int
 }
@@ -35,7 +35,7 @@ const (
 	lexSpace
 )
 
-func fill(in *Input, need int) int {
+func fill(in *Input, need uint) int {
 	// End of input has already been reached, nothing to do.
 	if in.eof {
 		fmt.Println("fill error: unexpected EOF")
@@ -59,8 +59,8 @@ func fill(in *Input, need int) int {
 
 	// Read new data (as much as possible to fill the buffer).
 	n, _ := in.file.Read(in.data[in.limit:SIZE])
-	in.limit += n
-	fmt.Printf("fill(%d): %v '%s'\n", need, in.data[:in.limit+1],
+	in.limit += uint(n)
+	fmt.Printf("fill(%u): %v '%s'\n", need, in.data[:in.limit+1],
 		string(in.data[:in.limit]))
 
 	// If read less than expected, this is the end of input.
@@ -69,7 +69,7 @@ func fill(in *Input, need int) int {
 	// If end of input, add padding so that the lexer can read
 	// the remaining characters at the end of buffer.
 	if in.eof {
-		for i := 0; i < YYMAXFILL; i += 1 {
+		for i := uint(0); i < YYMAXFILL; i += 1 {
 			in.data[in.limit+i] = 0
 		}
 		in.limit += YYMAXFILL
@@ -78,7 +78,7 @@ func fill(in *Input, need int) int {
 	return lexResume
 }
 
-func Lex(in *Input) (int, int) {
+func Lex(in *Input) (int, uint) {
 	var yych YYCTYPE
 
 	
@@ -237,7 +237,7 @@ func test(data string) int {
 		eof:    false,
 	}
 
-	need := 1
+	need := uint(1)
 	result := lexNeedMoreInput
 loop:
 	for {
